requirements.txt
flask==2.3.2
pymongo==4.3.3
python-dotenv==1.0.0
werkzeug==2.3.6
pyjwt==2.7.0
flask-cors==3.0.10